[REFACTOR][SCRIPT] Move IR script printer/builder into src/ir/script/
Today the centralized src/script/ subtree is the only place where every
dialect's IR types are referenced from script code, making it a sink that
imports IR, TIRX, and Relax. To enable adding new dialects without editing
src/script/, this commit (1 of 4) relocates the IR-layer pieces into
src/ir/script/{printer,builder}/ next to the rest of the IR layer. The
shared core (Doc IR, IRDocsifier dispatch, doc_printer, ir_builder/base)
stays at src/script/ as dialect-agnostic infrastructure.

Subsequent commits move TIRX, Relax, and finally tighten the src/script/
glob to dialect-free core; together they give every dialect ownership of
its own script subtree.

Notable details:

- Public IR-builder headers move to include/tvm/ir/script/builder/.
- Python: parser and builder modules move to python/tvm/ir/script/ with
  backward-compat shim packages left at the old paths so existing imports
  (`from tvm.script import ir`, `from tvm.script.ir_builder import ir as I`)
  keep working.
- src/ir/script/printer/distributed.cc dropped two unused includes
  (tvm/relax/distributed/global_info.h and ../relax/utils.h) that the
  original file carried but never referenced.
- src/ir/script/builder/ir.cc retains its includes of tvm/relax/analysis.h,
  tvm/tirx/function.h, and tvm/tirx/op.h; DeclFunction legitimately
  inspects whether a registered function is a relax::Function or a
  tirx::PrimFunc to derive the relax::FuncStructInfo, so the IR layer is
  the right home for that polyglot logic.
- src/script/printer/relax/distributed.cc updates its `../ir/utils.h`
  include to `../../../ir/script/printer/utils.h` to follow the move;
  the cleanup commit will replace this relative path with a public utils
  header.
