+++ include/tvm/script/ir_builder/relax/frame.h
@@ -20,11 +20,11 @@
 #define TVM_SCRIPT_IR_BUILDER_RELAX_FRAME_H_
 
 #include <tvm/ffi/reflection/registry.h>
+#include <tvm/ir/script/builder/frame.h>
+#include <tvm/ir/script/builder/ir.h>
 #include <tvm/relax/block_builder.h>
 #include <tvm/relax/expr.h>
 #include <tvm/script/ir_builder/base.h>
-#include <tvm/script/ir_builder/ir/frame.h>
-#include <tvm/script/ir_builder/ir/ir.h>
 
 #include <utility>
