+++ python/tvm/ir/script/parser/__init__.py
@@ -1,3 +1,4 @@
+# isort: skip_file
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -14,7 +15,21 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Backward-compat shim for ``tvm.script.parser.ir`` — moved to ``tvm.ir.script.parser``."""
+"""The ir module parser"""
 
-from tvm.ir.script.parser import *  # pylint: disable=wildcard-import,redefined-builtin
-from tvm.ir.script.parser import ir_module, pyfunc
+from tvm.ir import Range
+from tvm.ir.script.builder import *  # pylint: disable=redefined-builtin
+from . import parser as _parser
+from .entry import ir_module, pyfunc
+
+
+__all__ = [
+    "Range",
+    "dummy_global_info",
+    "ir_module",
+    "lookup_vdevice",
+    "module_attrs",
+    "module_global_infos",
+    "pyfunc",
+    "vdevice",
+]
